# perf: fix issue in Object.Insert on existing key